Update FixinatorClient.cfc
line 397 : Updated to retry on a 408 request time-out since on retry the number of files will be cut in half which may alleviate the amount of processing and prevent the 408 issue.

line 417: Updated the splitting of the payload to split more evenly on a retry.  Was splitting 5 file payload into payloads of 4 files and 1 file prior.