Update FixinatorClient.cfc
Updated default ignored paths to properly filter the various code repository folders if it is in the base directory which it normally is.  Having a slash prior to the folder name will not result in the folder being filtered out if it is in the root folder because the arguments.baseDirectory path is removed before checking for the ignored path and thus that first slash is removed.